Confirmed solution of Problem 2

+++ intro-and-basic-syntax/exercise/Problem 02 Rectangle Area/Program.cs
@@ -8,7 +8,7 @@ namespace Rectangle_Area
         private static void Main(string[] args)
         {
             //We receive from the Console the length for side a -> a double;
-            int a = int.Parse(Console.ReadLine());
+            double a = double.Parse(Console.ReadLine());
             //We receive from the Console the length for side b -> a double;
             double b = double.Parse(Console.ReadLine());
             //We calculate rectangle surface area;
